Create Smaller-Sum.java
Its an amazing question because there are certain concepts which are used in they are binary search, prefixsum, mapping, arrays. The time complexity should not exceed more the 10^8 ~ 1sec so, the given constraints are 1 <= n <= 10^5, 0 <= arr[i] <= 10^9. If you try to solve in o(n^2) it takes 10^10 > 10^8 so need to solve in nlogn or o(n). So we have solved in nlogn time complexity. We can still optimise it using HashMap at o(n).
NOTE: Read the question properly.