Create Nearest_smaller_tower.java
This is a medium level problem which came in Atlasian company. Not an easy one. You have to use two stacks for tracking previous smaller element than the current one. If you don't have it just put -1. 
1.You have to calculate the distance left and right distances out of which you have to get minimum one.
2.If distances are same and elements are not same then get minimum element index.
3.Satisfying the condition 2 and the elements are same then get minimum index.
The out put should be indexes of the respective elements instead of elements which makes confusing and not easy.
Example:
input 1: 
7
3 8 9 7 4 5 3
output:
-1 0 3 4 6 6 -1
input 2:
5
4 8 3 5 3
output 2:
2 2 -1 2 -1